+++ src/gemseo_benchmark/scenario.py
@@ -87,6 +87,7 @@ class Scenario:
         plot_all_histories: bool = True,
         use_log_scale: bool = False,
         log_gemseo_to_file: bool = False,
+        directory_path: Path | None = None,
         plot_only_median: bool = False,
     ) -> Results:
         """Execute the benchmarking scenario.
@@ -117,6 +118,8 @@ class Scenario:
             use_log_scale: Whether to use a logarithmic scale on the value axis.
             log_gemseo_to_file: Whether to save the GEMSEO log to a file
                 next to the performance history file.
+            directory_path: The path to the directory where the report
+                will be generated.
             plot_only_median: Whether to plot only the median and no other centile.
 
         Returns:
@@ -144,6 +147,7 @@ class Scenario:
                 max_eval_number_per_group,
                 plot_all_histories,
                 use_log_scale,
+                directory_path,
                 plot_only_median,
             )
 
@@ -198,12 +202,12 @@ class Scenario:
         Returns:
             The path to the directory.
         """
-        path = self._outputs_path / name
-        if path.is_dir() and overwrite:
-            shutil.rmtree(path)
+        directory_path = self._outputs_path / name
+        if directory_path.is_dir() and overwrite:
+            shutil.rmtree(directory_path)
 
-        path.mkdir(exist_ok=not overwrite)
-        return path
+        directory_path.mkdir(exist_ok=not overwrite)
+        return directory_path
 
     def __generate_report(
         self,
@@ -215,6 +219,7 @@ class Scenario:
         max_eval_number_per_group: dict[str, int],
         plot_all_histories: bool = True,
         use_log_scale: bool = False,
+        directory_path: Path | None = None,
         plot_only_median: bool = False,
     ) -> None:
         """Generate the benchmarking report.
@@ -235,10 +240,14 @@ class Scenario:
                 for the group.
             plot_all_histories: Whether to plot all the performance histories.
             use_log_scale: Whether to use a logarithmic scale on the value axis.
+            directory_path: The path to the directory where the report
+                will be generated.
             plot_only_median: Whether to plot only the median and no other centile.
         """
         report = Report(
-            self.__get_report_path(),
+            self.__get_report_path()
+            if directory_path is None
+            else directory_path,  # TODO: check existence
             self._algorithms_configurations_groups,
             problems_groups,
             Results(self._results_path),